requirements.txt
asgiref==3.6.0
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5d968ni_yn/croot/certifi_1671487774636/work/certifi
Django==3.2.5
django-filter==22.1
djangorestframework==3.14.0
docutils==0.16
importlib-metadata==6.0.0
Markdown==3.4.1
mysqlclient==2.1.1
pyasn1==0.4.8
pytz==2022.7.1
rsa==4.7.2
s3transfer==0.6.0
sqlparse==0.4.3
zipp==3.12.0
